Move strings helpers in their own file
