+++ packages/core/src/helpers.ts
@@ -7,48 +7,6 @@ import { isInjectionFile } from '@dd/core/helpers/plugins';
 import fs from 'fs';
 import path from 'path';
 
-// Format a duration 0h 0m 0s 0ms
-export const formatDuration = (duration: number) => {
-    const days = Math.floor(duration / 1000 / 60 / 60 / 24);
-    const usedDuration = duration - days * 24 * 60 * 60 * 1000;
-    const d = new Date(usedDuration);
-    const hours = d.getUTCHours();
-    const minutes = d.getUTCMinutes();
-    const seconds = d.getUTCSeconds();
-    const milliseconds = d.getUTCMilliseconds();
-    const timeString =
-        `${days ? `${days}d ` : ''}${hours ? `${hours}h ` : ''}${minutes ? `${minutes}m ` : ''}${
-            seconds ? `${seconds}s` : ''
-        }`.trim();
-    // Split here so we can show 0ms in case we have a duration of 0.
-    return `${timeString}${!timeString || milliseconds ? ` ${milliseconds}ms` : ''}`.trim();
-};
-
-// Truncate a string to a certain length.
-// Placing a [...] placeholder in the middle.
-// "A way too long sentence could be truncated a bit." => "A way too[...]could be truncated a bit."
-export const truncateString = (
-    str: string,
-    maxLength: number = 60,
-    placeholder: string = '[...]',
-) => {
-    if (str.length <= maxLength) {
-        return str;
-    }
-
-    // We want to keep at the very least 4 characters.
-    const stringLength = Math.max(4, maxLength - placeholder.length);
-
-    // We want to keep most of the end of the string, hence the 10 chars top limit for left.
-    const leftStop = Math.min(10, Math.floor(stringLength / 2));
-    const rightStop = stringLength - leftStop;
-
-    return `${str.slice(0, leftStop)}${placeholder}${str.slice(-rightStop)}`;
-};
-
-let index = 0;
-export const getUniqueId = () => `${Date.now()}.${performance.now()}.${++index}`;
-
 // Will only prepend the cwd if not already there.
 export const getAbsolutePath = (cwd: string, filepath: string) => {
     if (isInjectionFile(filepath)) {
