+++ packages/plugins/error-tracking/src/sourcemaps/sender.ts
@@ -3,7 +3,7 @@
 // Copyright 2019-Present Datadog, Inc.
 
 import { doRequest, NB_RETRIES } from '@dd/core/helpers/request';
-import { formatDuration } from '@dd/core/helpers';
+import { formatDuration } from '@dd/core/helpers/strings';
 import type { Logger, GlobalContext } from '@dd/core/types';
 import { File } from 'buffer';
 import chalk from 'chalk';
