+++ packages/plugins/injection/src/esbuild.ts
@@ -5,7 +5,8 @@
 import { INJECTED_FILE } from '@dd/core/constants';
 import { getEsbuildEntries } from '@dd/core/helpers/bundlers';
 import { outputFile } from '@dd/core/helpers/fs';
-import { getAbsolutePath, getUniqueId } from '@dd/core/helpers';
+import { getUniqueId } from '@dd/core/helpers/strings';
+import { getAbsolutePath } from '@dd/core/helpers';
 import type { Logger, PluginOptions, GlobalContext, ResolvedEntry } from '@dd/core/types';
 import { InjectPosition } from '@dd/core/types';
 import fs from 'fs';
