+++ packages/plugins/injection/src/xpack.ts
@@ -4,7 +4,7 @@
 
 import { INJECTED_FILE } from '@dd/core/constants';
 import { outputFileSync, rmSync } from '@dd/core/helpers/fs';
-import { getUniqueId } from '@dd/core/helpers';
+import { getUniqueId } from '@dd/core/helpers/strings';
 import type { GlobalContext, Logger, PluginOptions, ToInjectItem } from '@dd/core/types';
 import { InjectPosition } from '@dd/core/types';
 import { createRequire } from 'module';
