+++ packages/plugins/telemetry/src/common/output/files.ts
@@ -3,7 +3,7 @@
 // Copyright 2019-Present Datadog, Inc.
 
 import { outputJson } from '@dd/core/helpers/fs';
-import { formatDuration } from '@dd/core/helpers';
+import { formatDuration } from '@dd/core/helpers/strings';
 import type { Logger } from '@dd/core/types';
 import type { MetricToSend, OutputOptions, Report } from '@dd/telemetry-plugin/types';
 import path from 'path';
